fix(config): 修正目标语言默认值为中文并处理历史遗留auto值
- 将默认目标语言从 auto 改为 zh-CN，避免设置页面下拉选项排除 auto 导致的空白显示
- 在配置加载时自动将历史遗留的 auto 或空值转换为默认语言，并尝试保存更新
- 优化代码风格：简化闭包调用、移除冗余类型转换、统一错误处理
- 修复托盘图标回退逻辑，确保无默认图标时仍能显示内嵌图标
- 清理重复的引用传递，使用更简洁的API调用方式

+++ src-tauri/src/clipboard/paste.rs
@@ -74,7 +74,7 @@ fn trigger_paste_shortcut() {
             let mut inputs: [INPUT; 4] = std::mem::zeroed();
             // Ctrl down
             inputs[0].r#type = INPUT_TYPE(1);
-            inputs[0].Anonymous.ki.wVk = VIRTUAL_KEY(VK_CONTROL.0 as u16);
+            inputs[0].Anonymous.ki.wVk = VIRTUAL_KEY(VK_CONTROL.0);
             inputs[0].Anonymous.ki.dwFlags = KEYBD_EVENT_FLAGS(0);
             // V down
             inputs[1].r#type = INPUT_TYPE(1);
@@ -86,7 +86,7 @@ fn trigger_paste_shortcut() {
             inputs[2].Anonymous.ki.dwFlags = KEYBD_EVENT_FLAGS(KEYEVENTF_KEYUP);
             // Ctrl up
             inputs[3].r#type = INPUT_TYPE(1);
-            inputs[3].Anonymous.ki.wVk = VIRTUAL_KEY(VK_CONTROL.0 as u16);
+            inputs[3].Anonymous.ki.wVk = VIRTUAL_KEY(VK_CONTROL.0);
             inputs[3].Anonymous.ki.dwFlags = KEYBD_EVENT_FLAGS(KEYEVENTF_KEYUP);
             SendInput(&inputs, std::mem::size_of::<INPUT>() as i32);
         }
@@ -270,7 +270,7 @@ fn write_windows_clipboard_files(paths: &[String]) -> bool {
 }
 
 #[cfg(target_os = "windows")]
-#[allow(non_snake_case)]
+#[allow(non_snake_case, clippy::upper_case_acronyms)]
 #[repr(C)]
 struct DROPFILES {
     pFiles: u32,
